Numpy improvements

+++ mats_gym/envs/space_utils.py
@@ -6,7 +6,7 @@ def get_vehicle_action_space():
     return gymnasium.spaces.Box(
         low=np.array([0.0, -1.0, 0.0]),
         high=np.array([1.0, 1.0, 1.0]),
-        dtype=np.float64,
+        dtype=np.float32,
     )
 
 
@@ -16,7 +16,7 @@ def get_walker_action_space():
             "direction": gymnasium.spaces.Box(
                 low=np.array([-np.inf, -np.inf, -np.inf]),
                 high=np.array([np.inf, np.inf, np.inf]),
-                dtype=np.float64,
+                dtype=np.float32,
             ),
             "speed": gymnasium.spaces.Box(0, 12.5, shape=(1,), dtype=np.float32),
             "jump": gymnasium.spaces.Discrete(2),
